template/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "{{ project_name }}"
version = "0.1.0"
requires-python = ">=3.11"
authors = [{ name = "{{ author_name }}" }]
dependencies = [
    # Data
    "numpy",
    "pandas",
    "polars",
    "scikit-learn",
    "pyarrow",

    # Visualization
    "matplotlib",
    "seaborn",

    # GBDT
    "lightgbm",
    "xgboost",
    "catboost",

    # Deep Learning
    "torch",
    "torchvision",
    "torchaudio",
    "lightning",
    "transformers",

    # Uncomment for graph neural networks (requires matching torch + CUDA version):
    # "torch-geometric",

    # Uncomment for medical image segmentation:
    # "nnunetv2",

    # Configuration
    "pydantic>=2.0",
    "pyyaml",

    # Experiment tracking
    "wandb",

    # Utilities
    "tqdm",
    "click",
    "kaggle",
    "psutil",
    "rich",
]

[dependency-groups]
dev = [
    "ruff",
    "ty",
    "pytest",
    "pre-commit",
    "ipykernel",
    "jupyterlab",
]

[tool.hatch.build.targets.wheel]
packages = ["utils", "experiments", "tools"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.isort]
known-first-party = ["{{ package_name }}", "utils", "experiments"]

[tool.ty.environment]
python-version = "3.11"

[tool.ty.src]
exclude = ["tools/kaggle_login.py"]
